refactor: update `hover` API directory
