+++ packages/blocks/src/components/hover/index.ts
@@ -1,0 +1,2 @@
+export { HoverController } from './controller.js';
+export { whenHover } from './when-hover.js';
