+++ packages/blocks/src/components/index.ts
@@ -2,7 +2,7 @@ export * from './block-hub.js';
 export * from './block-selection.js';
 export * from './drag-indicator.js';
 export * from './file-drop-manager.js';
-export * from './hover-controller.js';
+export * from './hover/index.js';
 export * from './menu-divider.js';
 export * from './tooltip/index.js';
 import './portal.js';
