+++ packages/global/src/utils/index.ts
@@ -1,7 +1,6 @@
 export * from './assert.js';
 export * from './disposable.js';
 export * from './function.js';
-export * from './hover/index.js';
 export * from './iterable.js';
 export * from './slot.js';
 export * from './string.js';
